Repurpose getServiceTypes to get services
