--- src/profile/helpers/getServiceTypes.ts
@@ -1,20 +1,0 @@
-import {
-  ServiceConnectionsQuery,
-  ServiceType,
-} from '../../graphql/generatedTypes';
-
-export default function getServiceTypes(data?: ServiceConnectionsQuery) {
-  if (data?.myProfile?.serviceConnections) {
-    return data.myProfile.serviceConnections.edges
-      .map(edge => {
-        if (edge?.node?.service.type) {
-          return edge.node.service.type;
-        }
-        return false;
-      })
-      .filter((serviceType): serviceType is ServiceType =>
-        Boolean(serviceType)
-      );
-  }
-  return [];
-}
